+++ src/codegen/mod.rs
@@ -20047,6 +20047,7 @@ mod tests {
             create_set: Vec::new(),
             mutate_set: Vec::new(),
             write_intents: Vec::new(),
+            bounded_collection_ops: Vec::new(),
             blocks: Vec::new(),
         };
 
@@ -20420,6 +20421,7 @@ mod tests {
                     create_set: vec![],
                     mutate_set: vec![],
                     write_intents: vec![],
+                    bounded_collection_ops: vec![],
                     blocks: vec![IrBlock {
                         id: BlockId(0),
                         instructions: vec![],